backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
flask-socketio==5.3.2
tensorflow==2.16.2
pillow==10.4.0
numpy==1.26.4